app/requirements.txt
beautifulsoup4==4.12.2
dirtyjson==1.0.8
diskcache==5.4.0
openai==0.27.3
psycopg2==2.9.6
python_Levenshtein==0.21.1
regex==2023.3.23
requests==2.28.2
SQLAlchemy==2.0.15
